Moved analysis to App.vue; trying to work out passing it when one of the nav links is clicked
